refactor: extract indexer progress helpers under the length caps (#1826)
reindex_with_progress and reindex_audiobooks_with_progress grew past the
80-line soft cap as the verbose scan-progress feature (PRs #1806/#1808)
added inline progress-reporting closures, and indexer/mod.rs crossed the
800-line file cap in the same batch.

- Extract the per-phase progress-reporting closures into shared
  report_parse_progress/report_sync_progress helpers, and split each
  function's Phase-A stat+diff+tally logic into its own async helper
  (diff_ebook_library_for_reindex / diff_audiobook_library_for_reindex),
  bringing both functions to <=80 lines with no behavior change.
- Move the ScanUpdate wire shape, phase constants, and display/tally
  helpers into a new indexer/progress.rs; move the ebook reindex/
  reindex_with_progress entry points into a new indexer/ebooks.rs
  (mirroring the existing indexer/audiobooks.rs split), bringing
  indexer/mod.rs from 847 to 629 lines.
- Strip the four `(#1802)` issue-number parentheticals left in test
  comments, per the style guide's no-roadmap-refs-in-source rule.
- Update docs/architecture.md's indexer module map for the two new
  submodules.
