+++ db/src/epub_rewrite/tests.rs
@@ -841,9 +841,9 @@ async fn rewrite_all_epubs_with_overrides_rewrites_only_books_with_active_overri
     assert!(summary.errors.is_empty(), "{summary:?}");
 
     // One progress call per override row (A and C — B has no override), each
-    // naming the book being baked (#1802). The metadata fetch merges
-    // overrides, so A reports its overridden title. Order follows the
-    // override listing, so only membership is asserted.
+    // naming the book being baked. The metadata fetch merges overrides, so
+    // A reports its overridden title. Order follows the override listing,
+    // so only membership is asserted.
     let calls = progress.into_inner().unwrap();
     assert_eq!(calls.len(), 2, "{calls:?}");
     assert!(calls.iter().all(|(_, total, _)| *total == 2), "{calls:?}");
