+++ db/src/indexer/audiobooks.rs
@@ -7,9 +7,9 @@ use std::path::PathBuf;
 use sqlx::SqlitePool;
 
 use super::{
-    check_mass_missing, diff_library, diff_tallies, display_item, enumeration_is_trustworthy,
-    gc_missing_files_best_effort, root_display_name, ReindexStats, ScanUpdate, PHASE_PARSING,
-    PHASE_SYNCING, PHASE_WALKING,
+    check_mass_missing, diff_library, diff_tallies, enumeration_is_trustworthy,
+    gc_missing_files_best_effort, report_parse_progress, report_sync_progress, root_display_name,
+    ReindexDiff, ReindexStats, ScanUpdate, PHASE_WALKING,
 };
 use crate::{audiobook, books, sync};
 
@@ -79,21 +79,28 @@ fn project_groups_to_stat(groups: &[audiobook::AudiobookGroup]) -> Vec<crate::eb
         .collect()
 }
 
-/// [`reindex_audiobooks`] variant that reports verbose [`ScanUpdate`]
-/// events, mirroring [`super::reindex_with_progress`]: a [`PHASE_WALKING`]
-/// event before the walk, per-group [`PHASE_PARSING`] events during the
-/// tag read, and per-book [`PHASE_SYNCING`] events inside
-/// `sync_audiobooks` — parse and sync events carry the diff's
-/// [`omnibus_shared::ScanTallies`] and the current group's display path.
-/// Returns the scan's ghost-count tallies (issue #1057) so the caller can
-/// decide whether to attach a warning.
-pub async fn reindex_audiobooks_with_progress(
+/// Phase-A output of [`reindex_audiobooks_with_progress`]: the grouped
+/// audiobooks plus the walk's diff, breaker counts, and progress tallies —
+/// mirrors the ebook pipeline's own Phase-A aggregate in
+/// [`super::reindex_with_progress`].
+struct AudiobookDiffPhase {
+    groups: Vec<audiobook::AudiobookGroup>,
+    diff: ReindexDiff,
+    removed_count: usize,
+    moved_count: usize,
+    db_file_backed: usize,
+    tallies: omnibus_shared::ScanTallies,
+    root_name: String,
+}
+
+/// Phase A of [`reindex_audiobooks_with_progress`]: group the library,
+/// diff against the DB, and derive the walk's fixed breaker counts and
+/// progress tallies up front so every later parse/sync event reports the
+/// same numbers.
+async fn diff_audiobook_library_for_reindex(
     pool: &SqlitePool,
     library_path: &str,
-    on_progress: impl FnMut(ScanUpdate) + Send + 'static,
-) -> anyhow::Result<ReindexStats> {
-    let mut on_progress = on_progress;
-    on_progress(ScanUpdate::phase(PHASE_WALKING));
+) -> anyhow::Result<AudiobookDiffPhase> {
     let (groups, signals) = stat_and_group_audiobooks(library_path).await?;
 
     // Diff groups against DB rows (project groups to the ebook StatEntry shape
@@ -141,42 +148,83 @@ pub async fn reindex_audiobooks_with_progress(
     let tallies = diff_tallies(groups_as_stat.len(), &diff);
     let root_name = root_display_name(library_path);
 
-    // Phase B: parse only the New and Changed groups.
+    Ok(AudiobookDiffPhase {
+        groups,
+        diff,
+        removed_count,
+        moved_count,
+        db_file_backed,
+        tallies,
+        root_name,
+    })
+}
+
+/// Resolve the diff's New/Changed filename buckets back to full
+/// [`audiobook::AudiobookGroup`]s for Phase-B parsing. `groups` is consumed
+/// (moved into a lookup map) since Phase A has no further use for it.
+fn select_parse_groups(
+    groups: Vec<audiobook::AudiobookGroup>,
+    diff: &ReindexDiff,
+) -> (
+    Vec<audiobook::AudiobookGroup>,
+    Vec<audiobook::AudiobookGroup>,
+) {
     let groups_by_group_path: std::collections::HashMap<String, audiobook::AudiobookGroup> = groups
         .into_iter()
         .filter(|g| !g.scan_key.is_empty())
         .map(|g| (g.group_path.clone(), g))
         .collect();
 
-    let new_groups: Vec<audiobook::AudiobookGroup> = diff
+    let new_groups = diff
         .new
         .iter()
         .filter_map(|t| groups_by_group_path.get(&t.filename).cloned())
         .collect();
-    let changed_groups: Vec<audiobook::AudiobookGroup> = diff
+    let changed_groups = diff
         .changed
         .iter()
         .filter_map(|t| groups_by_group_path.get(&t.filename).cloned())
         .collect();
+    (new_groups, changed_groups)
+}
+
+/// [`reindex_audiobooks`] variant that reports verbose [`ScanUpdate`]
+/// events, mirroring [`super::reindex_with_progress`]: a [`PHASE_WALKING`]
+/// event before the walk, per-group parse events during the tag read, and
+/// per-book sync events inside `sync_audiobooks` — parse and sync events
+/// carry the diff's [`omnibus_shared::ScanTallies`] and the current
+/// group's display path. Returns the scan's ghost-count tallies (issue
+/// #1057) so the caller can decide whether to attach a warning.
+pub async fn reindex_audiobooks_with_progress(
+    pool: &SqlitePool,
+    library_path: &str,
+    on_progress: impl FnMut(ScanUpdate) + Send + 'static,
+) -> anyhow::Result<ReindexStats> {
+    let mut on_progress = on_progress;
+    on_progress(ScanUpdate::phase(PHASE_WALKING));
+
+    let phase_a = diff_audiobook_library_for_reindex(pool, library_path).await?;
+
+    // Phase B: parse only the New and Changed groups.
+    let (new_groups, changed_groups) = select_parse_groups(phase_a.groups, &phase_a.diff);
 
     let parse_total = u32::try_from(new_groups.len() + changed_groups.len()).unwrap_or(u32::MAX);
-    let root_for_parse = library_root.clone();
-    let root_name_for_parse = root_name.clone();
+    let root_for_parse = PathBuf::from(library_path);
+    let root_name_for_parse = phase_a.root_name.clone();
+    let tallies = phase_a.tallies;
     // The callback moves into the blocking task for per-group reporting and
     // rides back out in the return value for the sync phase below.
     let (parsed, on_progress) = tokio::task::spawn_blocking(move || {
         let mut parsed_count = 0u32;
         let mut report = |g: &audiobook::AudiobookGroup| {
             parsed_count = parsed_count.saturating_add(1);
-            on_progress(ScanUpdate {
-                processed: parsed_count,
-                total: Some(parse_total),
-                detail: omnibus_shared::TaskDetail {
-                    phase: Some(PHASE_PARSING.to_string()),
-                    current_item: Some(display_item(&root_name_for_parse, &g.scan_key)),
-                    tallies: Some(tallies),
-                },
-            });
+            on_progress(report_parse_progress(
+                parsed_count,
+                parse_total,
+                &root_name_for_parse,
+                &g.scan_key,
+                tallies,
+            ));
         };
         let new_books =
             audiobook::parse_groups_with_progress(new_groups, &root_for_parse, &mut report);
@@ -190,26 +238,24 @@ pub async fn reindex_audiobooks_with_progress(
     let plan = sync::AudiobookSyncPlan {
         new_books: parsed.0,
         changed_books: parsed.1,
-        moved: diff.moved,
-        removed_uuids: diff.removed,
-        backfill: diff.backfill,
+        moved: phase_a.diff.moved,
+        removed_uuids: phase_a.diff.removed,
+        backfill: phase_a.diff.backfill,
     };
     sync::sync_audiobooks_with_progress(pool, library_path, plan, |processed, total, current| {
-        on_progress(ScanUpdate {
+        on_progress(report_sync_progress(
             processed,
-            total: Some(total),
-            detail: omnibus_shared::TaskDetail {
-                phase: Some(PHASE_SYNCING.to_string()),
-                current_item: current.map(|c| display_item(&root_name, c)),
-                tallies: Some(tallies),
-            },
-        });
+            total,
+            &phase_a.root_name,
+            current,
+            tallies,
+        ));
     })
     .await?;
     gc_missing_files_best_effort(pool).await;
     Ok(ReindexStats {
-        removed: removed_count,
-        file_backed_total: db_file_backed,
-        moved: moved_count,
+        removed: phase_a.removed_count,
+        file_backed_total: phase_a.db_file_backed,
+        moved: phase_a.moved_count,
     })
 }
